+++ inst/shiny/shinytest2.R
@@ -1,9 +1,0 @@
-library(shinytest2)
-
-
-# shiny_app <- SMART::run_smart()
-# app <- AppDriver$new(shiny_app, name = "SMART")
-record_test('inst/shiny/')
-
-
-