+++ tests/testthat/test-plot_hist.R
@@ -15,4 +15,17 @@ test_that("Check plot_hist function works as expected", {
 
 })
 
-
+test_that("{shinytest2} recording: e2e_plot_hist", {
+  app <- shinytest2::AppDriver$new(app_dir = '../../inst/shiny', name = "e2e_plot_hist")
+  app$set_inputs(tabs = "select")
+  app$set_inputs(selectSel = "select_user")
+  app$upload_file(`select_user-ras` = path)
+  app$set_inputs(`select_user-name` = "bio")
+  app$click("select_user-run")
+  app$set_inputs(tabs = "plot")
+  app$set_inputs(plotSel = "plot_hist")
+  app$set_inputs(`plot_hist-pal` = "YlOrRd")
+  app$click("plot_hist-run")
+  common <- app$get_value(export = "common")
+  expect_is(common$hist, 'histogram')
+})