+++ tests/testthat/test-plot_scatter.R
@@ -17,4 +17,16 @@ test_that("Check plot_scatter function works as expected", {
 
 })
 
-
+test_that("{shinytest2} recording: e2e_plot_scatter", {
+  app <- shinytest2::AppDriver$new(app_dir = '../../inst/shiny', name = "e2e_plot_scatter")
+  app$set_inputs(tabs = "select")
+  app$set_inputs(selectSel = "select_user")
+  app$upload_file(`select_user-ras` = path)
+  app$set_inputs(`select_user-name` = "bio")
+  app$click("select_user-run")
+  app$set_inputs(tabs = "plot")
+  app$set_inputs(plotSel = "plot_scatter")
+  app$click("plot_scatter-run")
+  common <- app$get_value(export = "common")
+  expect_is(common$scat, 'data.frame')
+})